Modificaciones menores para mejor legibilidad
